+++ report/src/App.vue
@@ -1,86 +1,90 @@
 <template>
-  <div class="bg-gray-50 min-h-screen">
-      <loading :active="$root.isLoading"></loading>
-      <div class="flex flex-wrap flex-col items-center">
+  <div class="bg-gray-50 min-h-screen p-4">
+    <loading :active="$root.isLoading"></loading>
+    <div class="flex flex-wrap flex-col items-center">
       <div class="w-full lg:w-8/12 my-6">
-        <div class="flex justify-center my-6"><img src="@/assets/logo.png" class="mx-auto w-40 h-40" alt="WAG Media" title="WAG Media" /></div>
+        <div class="flex justify-center my-6"><img src="@/assets/logo.png" class="mx-auto w-40 h-40" alt="WAG Media"
+            title="WAG Media" /></div>
         <h2 class="text-xl font-semibold">Valuated Messages</h2>
-        <div class="p-5 mt-6 w-full h-80">
-          <div class="flex w-full justify-between">
-            <div class="w-5/12"><canvas id="graph-transaction-canvas"></canvas></div>
-            <div class="w-5/12"><canvas id="graph-value-canvas"></canvas></div>
+        <div class="p-5 mt-6 w-full">
+          <div class="flex flex-col lg:flex-row w-full justify-between">
+            <div class="w-full lg:w-5/12 my-4"><canvas id="graph-transaction-canvas h-80"></canvas></div>
+            <div class="w-full lg:w-5/12 my-4"><canvas id="graph-value-canvas h-80"></canvas></div>
           </div>
         </div>
         <div class="px-4 border rounded-lg bg-white border-gray-300 dark:border-gray-700 p-5 mt-6">
           <div class="flex flex-wrap mb-6 justify-between">
-            <div class="w-3/12">
+            <div class="w-full lg:w-3/12 mb-4">
               <label for="treasuryId" class="block mb-2">Treasury:</label>
-              <multiselect v-model="searchFilter.treasuryId" :allow-empty="false" :showLabels="false" :showPointer="false" label="name" track-by="id" :options="treasuries" @input="$refs.vuetable.reload()">
-                <template slot="singleLabel" slot-scope="{ option }"><div>{{option.name}}</div></template>
-                <template slot="option" slot-scope="{ option }"><div>{{option.name}}</div></template>
+              <multiselect v-model="searchFilter.treasuryId" :allow-empty="false" :showLabels="false"
+                :showPointer="false" label="name" track-by="id" :options="treasuries">
+                <template slot="singleLabel" slot-scope="{ option }">
+                  <div>{{ option.name }}</div>
+                </template>
+                <template slot="option" slot-scope="{ option }">
+                  <div>{{ option.name }}</div>
+                </template>
               </multiselect>
             </div>
-            <div class="w-3/12">
+            <div class="w-full lg:w-3/12 mb-4">
               <label for="status" class="block mb-2">Status:</label>
-              <multiselect v-model="searchFilter.status" :allow-empty="false" :showLabels="false" :showPointer="false" label="label" track-by="value" :options="statusOptions" @input="$refs.vuetable.reload()">
-                <template slot="singleLabel" slot-scope="{ option }"><div>{{option.label}}</div></template>
-                <template slot="option" slot-scope="{ option }"><div>{{option.label}}</div></template>
+              <multiselect v-model="searchFilter.status" :allow-empty="false" :showLabels="false" :showPointer="false"
+                label="label" track-by="value" :options="statusOptions">
+                <template slot="singleLabel" slot-scope="{ option }">
+                  <div>{{ option.label }}</div>
+                </template>
+                <template slot="option" slot-scope="{ option }">
+                  <div>{{ option.label }}</div>
+                </template>
               </multiselect>
             </div>
-            <div class="w-3/12">
+            <div class="w-full lg:w-3/12">
               <label for="date" class="block mb-2">Date Range:</label>
               <date-picker v-model="datepicker" range format="MM/DD/YYYY"></date-picker>
             </div>
           </div>
-    <vuetable
-      ref="vuetable"
-      :api-url="apiUrl"
-      :fields="fields"
-      data-path="data"
-      pagination-path=""
-      :css="css.table"
-      :sort-order="sortOrder"
-      :per-page="20"
-      :query-params="makeQueryParams"
-      @vuetable:pagination-data="onPaginationData"
-      @vuetable:loading="onLoading"
-      @vuetable:loaded="onLoaded"
-      @vuetable:load-success="onLoadSuccess"
-      @vuetable:row-clicked="onRowClicked"
-      detail-row-component="detail-row-public"
-    >
-    <template slot="date" slot-scope="{ rowData }">
-        {{ getDate(rowData.timestamp) }}
-    </template>
-    <template slot="valuation" slot-scope="{ rowData }">
-        {{ rowData.value }} {{ rowData.coinName }}
-    </template>
-    <template slot="status" slot-scope="{ rowData }">
-      <div><ValuatedMessageStatusIcon :status="rowData.status" /></div>
-    </template>
-    </vuetable>
-    <div class="w-full flex items-center mt-6 px-6">
-      <vuetable-pagination-info :css="css.paginationInfo" ref="paginationInfo"
-      ></vuetable-pagination-info>
-      <vuetable-pagination
-        ref="pagination"
-        :css="css.pagination"
-        @vuetable-pagination:change-page="onChangePage"
-      ></vuetable-pagination>
-    </div>
-  </div>
+
+          <v-data-table :headers="headers" :items="valuations" :page.sync="page" :options.sync="options"
+            :server-items-length="totalValuations" :loading="this.$root.isLoading" show-expand :mobile-breakpoint="0"
+            hide-default-footer :single-expand="false" :expanded.sync="expanded" :items-per-page="10"
+            class="elevation-1" @page-count="pageCount = $event">
+            <template v-slot:expanded-item="{ headers, item }">
+              <td :colspan="headers.length">
+                <ValuatedMessageDetail :item="item" />
+              </td>
+            </template>
+            <template v-slot:item.valuation="{ item }">
+              {{ item.value }} {{ item.coinName }}
+            </template>
+
+            <template v-slot:item.date="{ item }">
+              {{ getDate(item.timestamp) }}
+            </template>
+
+            <template v-slot:item.status="{ item }">
+              <div>
+                <ValuatedMessageStatusIcon :status="item.status" />
+              </div>
+            </template>
+          </v-data-table>
+
+          <div class="flex flex-wrap w-full items-center mt-6">
+            <div class="w-1/4 pr-4 text-sm whitespace-nowrap">{{ totalValuations == 1 ? totalValuations + ' Valuation' : totalValuations + ' Valuations' }}</div>
+            <div class="w-3/4 flex justify-end">
+              <v-pagination v-model="page" :length="pageCount" :totalVisible="10"></v-pagination>
+            </div>
+          </div>
+        </div>
       </div>
-  </div>
+    </div>
+    </div>
+
 </template>
 <script>
 import Loading from "vue-loading-overlay";
 import "vue-loading-overlay/dist/vue-loading.css";
 import moment from "moment";
-import Vue from "vue";
 import Chart from "chart.js/auto";
-import Vuetable from "vuetable-2/src/components/Vuetable";
-import VuetablePagination from "vuetable-2/src/components/VuetablePagination";
-import VuetablePaginationInfo from "vuetable-2/src/components/VuetablePaginationInfo";
 import ValuatedMessageStatusIcon from "@/components/ValuatedMessageStatusIcon";
 import ValuatedMessageDetail from "@/components/ValuatedMessageDetail";
 import { interpolatePlasma } from "d3-scale-chromatic";
@@ -89,15 +93,10 @@ import "vue2-datepicker/index.css";
 import API from "@/services/api";
 import Multiselect from "vue-multiselect";
 
-Vue.component("detail-row-public", ValuatedMessageDetail);
-
 export default {
   name: "App",
   components: {
     DatePicker,
-    Vuetable,
-    VuetablePagination,
-    VuetablePaginationInfo,
     ValuatedMessageDetail,
     ValuatedMessageStatusIcon,
     Multiselect,
@@ -119,7 +118,7 @@ export default {
       }
     );
 
-    API.request("treasury/public")
+    API.request("/treasuries/")
       .then((response) => {
         this.treasuries = response.data;
         this.treasuries.unshift({ id: null, name: "All" });
@@ -129,12 +128,18 @@ export default {
         this.treasuriesLoaded = false;
         console.log(error);
       });
+
+    this.getData().then(response => this.renderGraphs(response.data.stats));
   },
   data() {
     return {
       transactionGraph: null,
       valueGraph: null,
       datepicker: null,
+      options: {},
+      page: 1,
+      pageCount: 0,
+      expanded: [],
       searchFilter: {
         dateStart: null,
         dateEnd: null,
@@ -158,78 +163,47 @@ export default {
           label: "Other",
         },
       ],
+      headers: [
+        {
+          text: "Source",
+          value: "source",
+          sortField: "valuation.source",
+          sortable: true,
+          align: "start",
+        },
+        {
+          text: "Valuation",
+          value: "valuation",
+          sortable: false,
+          align: "start",
+        },
+        {
+          text: "Treasury",
+          value: "name",
+          sortField: "treasury.name",
+          sortable: true,
+          align: "start",
+        },
+        {
+          text: "Date",
+          value: "date",
+          sortField: "valuation.timestamp",
+          sortable: true,
+          align: "start",
+        },
+        {
+          text: "Status",
+          value: "status",
+          sortField: "valuation.status",
+          sortable: true,
+          align: "start",
+        },
+      ],
       treasuriesLoaded: false,
       dataTableLoaded: false,
       treasuries: [],
-      oldQueryParams: null,
-      currentQueryParams: null,
-      rerenderGraphs: true,
-      apiUrl: process.env.VUE_APP_API_URL + "valuation/public",
-      fields: [
-        {
-          name: "source",
-          title: "Source",
-          sortField: "valuation.source",
-        },
-        {
-          name: "__slot:valuation",
-          title: "Valuation",
-        },
-        {
-          name: "name",
-          title: "Treasury",
-          sortField: "treasury.name",
-        },
-        {
-          name: "__slot:date",
-          title: "Date",
-          sortField: "valuation.timestamp",
-        },
-        {
-          name: "__slot:status",
-          title: "Status",
-          sortField: "valuation.status",
-        },
-      ],
-      sortOrder: [{ field: "valuation.timestamp", direction: "desc" }],
-      css: {
-        table: {
-          tableWrapper: "w-full",
-          tableHeaderClass: "mb-0",
-          tableBodyClass: "mb-0",
-          tableClass: "w-full",
-          loadingClass: "loading",
-          ascendingIcon: "fa fa-chevron-up",
-          descendingIcon: "fa fa-chevron-down",
-          ascendingClass: "sorted-asc",
-          descendingClass: "sorted-desc",
-          sortableIcon: "fa fa-sort",
-          detailRowClass: "vuetable-detail-row",
-          handleIcon: "fa fa-bars text-secondary",
-          renderIcon(classes) {
-            return `<i class="${classes.join(" ")}"></span>`;
-          },
-        },
-        pagination: {
-          wrapperClass: "pagination ml-auto",
-          activeClass: "active",
-          disabledClass: "disabled",
-          pageClass: "page-item",
-          linkClass: "page-link",
-          paginationClass: "pagination",
-          paginationInfoClass: "float-left",
-          dropdownClass: "form-control",
-          icons: {
-            first: "fa fa-backward-fast",
-            prev: "fa fa-chevron-left",
-            next: "fa fa-chevron-right",
-            last: "fa fa-forward-fast",
-          },
-        },
-        paginationInfo: {
-          infoClass: "text-sm",
-        },
-      },
+      valuations: [],
+      totalValuations: 0
     };
   },
   watch: {
@@ -240,31 +214,26 @@ export default {
       this.searchFilter.dateEnd = Math.floor(
         this.datepicker[1]?.getTime() / 1000
       );
-      this.$refs.vuetable.reload();
+    },
+
+    searchFilter: {
+      handler() {
+        this.page = 1
+        this.getData().then((response) => this.renderGraphs(response.data.stats));
+      },
+      deep: true,
+    },
+
+    options: {
+      handler() {
+        this.getData();
+      },
+      deep: true,
     },
   },
   methods: {
     getDate(timestamp) {
       return moment.unix(timestamp).format("MM/DD/YYYY HH:mm");
-    },
-    onPaginationData(paginationData) {
-      this.$refs.pagination.setPaginationData(paginationData);
-      this.$refs.paginationInfo.setPaginationData(paginationData);
-    },
-    onChangePage(page) {
-      this.$refs.vuetable.changePage(page);
-    },
-    onLoading() {
-      this.dataTableLoaded = false;
-    },
-    onLoaded() {
-      this.dataTableLoaded = true;
-    },
-    onLoadSuccess(data) {
-      this.renderGraphs(data.data.stats);
-    },
-    onRowClicked(data) {
-      this.$refs.vuetable.toggleDetailRow(data.id);
     },
     renderGraphs(data) {
       if (data.length == 0) {
@@ -311,13 +280,7 @@ export default {
                   callbacks: {
                     label: function (tooltipItem) {
                       const numberFormatter = Intl.NumberFormat("en-US");
-                      return (
-                        tooltipItem.label +
-                        ": " +
-                        numberFormatter.format(
-                          tooltipItem.dataset.data[tooltipItem.dataIndex]
-                        )
-                      );
+                      return (tooltipItem.label +  ": " + numberFormatter.format(tooltipItem.dataset.data[tooltipItem.dataIndex]));
                     },
                   },
                 },
@@ -353,13 +316,7 @@ export default {
                   callbacks: {
                     label: function (tooltipItem) {
                       const numberFormatter = Intl.NumberFormat("en-US");
-                      return (
-                        tooltipItem.label +
-                        ": " +
-                        numberFormatter.format(
-                          tooltipItem.dataset.data[tooltipItem.dataIndex]
-                        )
-                      );
+                      return (tooltipItem.label + ": " + numberFormatter.format(tooltipItem.dataset.data[tooltipItem.dataIndex]));
                     },
                   },
                 },
@@ -376,19 +333,51 @@ export default {
         );
       }
     },
-    makeQueryParams(sortOrder, currentPage, perPage) {
+    makeQueryParams() {
+      let options = {
+        paginated: true,
+      };
+
       let searchFilter = { ...this.searchFilter };
       searchFilter.treasuryId = this.searchFilter.treasuryId?.id;
       searchFilter.status = this.searchFilter.status?.value;
 
-      return {
-        searchFilter: searchFilter,
-        paginated: true,
-        sortField: sortOrder[0].sortField,
-        sortOrder: sortOrder[0].direction,
-        pageNo: currentPage,
-        pageSize: perPage,
-      };
+      options.searchFilter = searchFilter;
+      if (this.options.sortBy) {
+        if (this.options.sortBy.length === 0) {
+          options.sortField = "valuation.timestamp";
+        } else {
+          options.sortField = this.headers.filter((e) => e.value === this.options.sortBy[0])[0].sortField || this.options.sortBy[0];
+        }
+      } else {
+        options.sortField = "valuation.timestamp";
+      }
+
+      if (!this.options.sortDesc || this.options.sortDesc.length === 0) {
+        options.sortOrder = "DESC";
+      } else {
+        options.sortOrder = this.options.sortDesc[0] ? "DESC" : "ASC";
+      }
+
+      options.pageSize = 10;
+      options.pageNo = this.page;
+
+      return options;
+    },
+    getData() {
+      return API.request("/valuations/", this.makeQueryParams())
+        .then((response) => {
+          this.dataTableLoaded = true;
+          this.valuations = response.data.data;
+          this.totalValuations = response.data.total;
+          this.pageCount = response.data.last_page;
+
+          return response;
+        })
+        .catch((e) => {
+          this.dataTableLoaded = false;
+          console.log(e);
+        });
     },
   },
 };